Improve file structure; Put the React integration example in /demo
